Add conditional check to font install

+++ fonts/install.sh
@@ -4,6 +4,8 @@
 #
 # Commands to run during installation 
 
-git clone https://github.com/powerline/fonts.git "$TMPDIR/powerline-fonts"
-find "$TMPDIR/powerline-fonts/SourceCodePro" -type f -name "*.otf" -exec cp {} ~/Library/Fonts \;
-rm -rf "$TMPDIR/powerline-fonts"
+if [ $(find ~/Library/Fonts -name "Source Code Pro*" | wc -l) -eq 0 ]; then
+  git clone https://github.com/powerline/fonts.git "$TMPDIR/powerline-fonts"
+  find "$TMPDIR/powerline-fonts/SourceCodePro" -type f -name "*.otf" -exec cp {} ~/Library/Fonts \;
+  rm -rf "$TMPDIR/powerline-fonts"
+fi
